+++ Gherkin/common.py
@@ -1,8 +1,4 @@
 import re
-import chromadb
-from chromadb.config import Settings
-import common_database
-import common_embed
 
 ### parse model@host
 
@@ -120,79 +116,3 @@ def get_internal_id_of_keyword(internal_description_id: str) -> str:
         raise ValueError("Not id of a description")
     return f"{get_external_id(internal_description_id)}-k"
 
-
-### database queries
-
-def search_keywords(db_path: str, host: str, model: str, project: str, keyword_type: str, keyword: str, nb_results:int) -> list[dict[str, str]]:
-    """
-    Extract the nearest neighbours of a keyword from a Chroma database.
-
-    Args:
-        db_path: The path to the Chroma database.
-        host: The host of the model.
-        model: The name of the model.
-        project: The name of the project.
-        keyword_type: The type of keyword to extract.
-        keyword: The keyword to search for.
-        nb_results: The number of results to return.
-
-    Returns:
-        A list of dictionaries with the following keys:
-            - id: The external id of the matching keyword.
-            - match: What is matching the searched keyword ('keyword' or 'description').
-            - keyword: The text of the matching keyword.
-            - description: The description of the matching keyword. Absent if the keyword has no description.
-            - keyword_distance: (May be missing if description matches) the distance between the matching keyword and the searched keyword.
-            - description_distance: (May be missing if keyword matches) the distance between the description of the matching keyword and the searched keyword.
-
-    Raises:
-        ValueError: If the model and/or project do not exist in the database.
-    """
-
-    model_id = common_database.get_model_id(db_path, model, host)
-    if not model_id:
-        raise ValueError(f"Model {model} at host {host} do not exist in database.")
-
-    # Initialize Chroma client
-    chroma_client = chromadb.PersistentClient(path=db_path, settings=Settings(anonymized_telemetry=False))
-
-    # Get the appropriate collection
-    collection_name = f"{get_collection_name(model_id, project, keyword_type)}"
-
-    embedding_function = common_embed.build_embedding_function(host, model)
-    try:
-        collection = chroma_client.get_collection(
-            name=collection_name,
-            embedding_function=embedding_function
-        )
-    except chromadb.errors.InvalidCollectionException as e:
-        raise ValueError(f"Error: Model {model} and/or project {project} do not exist in database {db_path}.") from e
-
-    documents = collection.get(include=['documents'])
-
-    # Perform the search query
-    search_results = collection.query(
-        query_texts=[keyword],
-        n_results=nb_results
-    )
-
-    data = []
-    result_ids = search_results['ids'][0]
-    result_docs = search_results['documents'][0]
-    result_dists = search_results['distances'][0]
-    for i in range(len(result_ids)):
-        if get_document_type(result_ids[i]) == 'keyword':
-            d = { 'id': get_external_id(result_ids[i]), 'match': 'keyword', 'keyword': result_docs[i], 'keyword_distance': result_dists[i]}
-            description_id = get_internal_id_of_description(result_ids[i])
-            if description_id in documents['ids']:
-                d['description'] = documents['documents'][documents['ids'].index(description_id)]
-                if description_id in result_ids:
-                    d['description_distance'] = result_dists[result_ids.index(description_id)]
-        else:
-            d = { 'id': get_external_id(result_ids[i]), 'match': 'description', 'description': result_docs[i], 'description_distance': result_dists[i]}
-            keyword_id = get_internal_id_of_keyword(result_ids[i])
-            d['keyword'] = documents['documents'][documents['ids'].index(keyword_id)]
-            if keyword_id in result_ids:
-                d['keyword_distance'] = result_dists[result_ids.index(keyword_id)]
-        data.append(d)
-    return data
