+++ Gherkin/fill_database.py
@@ -1,10 +1,9 @@
 import argparse
-import chromadb
-from chromadb.config import Settings
 import json
+
 import common
-import common_embed
-import common_database
+import model_db
+import vector_db
 
 def main():
     parser = argparse.ArgumentParser(description="Compute embedding vectors and store them in the Chroma database.")
@@ -15,25 +14,14 @@ def main():
     args = parser.parse_args()
     model, host = common.parse_model_and_host(args.model)
 
-    chroma_client = chromadb.PersistentClient(path=args.db_path, settings=Settings(anonymized_telemetry=False))
-    embedding_function = common_embed.build_embedding_function(host, model)
+    # Setup database
+    model_db.setup_database(args.db_path)
 
-    common_database.setup_database(args.db_path)
-    model_id = common_database.get_model_id(args.db_path, model, host)
-    if not model_id:
-        model_id = common_database.add_model_and_host(args.db_path, model, host)
-
+    # Load keywords
     with open(args.keyword_file, 'r', encoding='utf-8') as file:
         data = json.load(file)
 
-    for type in ["Context", "Action", "Outcome"]:
-        keywords = [item for item in data['keywords'] if item['type'] == type]
-        if keywords != []:
-            collection = chroma_client.get_or_create_collection(name=f"{common.get_collection_name(model_id, args.project, type)}", embedding_function=embedding_function)
-            collection.upsert(documents=[item['keyword'] for item in keywords], ids=[f"{item['id']}-k" for item in keywords])
-            documented_keywords = [item for item in keywords if len(item['description']) > 0]
-            if documented_keywords != []:
-                collection.upsert(documents=[item['description'] for item in documented_keywords], ids=[f"{item['id']}-d" for item in documented_keywords])
+    vector_db.fill_database(args.db_path, model, host, args.project, data)
 
 if __name__ == "__main__":
     main()
